requirements.txt
ipykernel
ipywidgets
kaggle
transformers
huggingface_hub
datasets
openai
pandas
polars
scikit-learn
sentence-transformers
python-dotenv
torch
torchvision
torchaudio
lightgbm>=4.4.0
autogluon
fairlearn
matplotlib
langdetect
cleanlab
anywidget
seaborn
cchardet
plotly
dython
openml
ray-tune>=2.10.0,<2.32
ray-default>=2.10.0,<2.32
optbinning
sliceline
openfe